% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CensTimeTwoArm.R
\name{CensTimeTwoArm}
\alias{CensTimeTwoArm}
\title{Title}
\usage{
CensTimeTwoArm(
  N.0,
  N.1,
  d,
  l,
  alpha0.t,
  nu0.t,
  alpha1.t,
  nu1.t,
  gamma.c,
  s,
  m,
  design2
)
}
\arguments{
\item{m}{}
}
\description{
Title
}
