% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumEventsSub.R
\name{NumEventsSub}
\alias{NumEventsSub}
\title{Title}
\usage{
NumEventsSub(
  N = NULL,
  d = NULL,
  l = NULL,
  gamma = NULL,
  s = NULL,
  m = NULL,
  alpha = NULL,
  nu = NULL,
  design1 = NULL
)
}
\arguments{
\item{design1}{}
}
\description{
Title
}
