% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumEventsSub.TwoArm.R
\name{NumEventsSub.TwoArm}
\alias{NumEventsSub.TwoArm}
\title{Title}
\usage{
NumEventsSub.TwoArm(
  N.0 = NULL,
  N.1 = NULL,
  ratio = NULL,
  l = NULL,
  gamma.c = NULL,
  alpha0.t = NULL,
  nu0.t = NULL,
  alpha1.t = NULL,
  nu1.t = NULL,
  s,
  m,
  design2 = NULL
)
}
\description{
Title
}
