% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumEventsSubTwoArm.R
\name{NumEventsSubTwoArm}
\alias{NumEventsSubTwoArm}
\title{Title}
\usage{
NumEventsSubTwoArm(
  N.0 = NULL,
  N.1 = NULL,
  ratio = NULL,
  d = NULL,
  alpha0.t = NULL,
  nu0.t = NULL,
  alpha1.t,
  nu1.t,
  gamma.c,
  s,
  m,
  l,
  design2 = NULL
)
}
\arguments{
\item{N.0}{example}

\item{N.1}{example}

\item{ratio}{example}

\item{d}{1}

\item{alpha0.t}{1}

\item{nu0.t}{1}

\item{alpha1.t}{1}

\item{nu1.t}{1}

\item{gamma.c}{1}

\item{s}{1}

\item{m}{1}

\item{l}{1}
}
\description{
Title
}
