% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObsTime.TwoArm.R
\name{ObsTime.TwoArm}
\alias{ObsTime.TwoArm}
\title{Title}
\usage{
ObsTime.TwoArm(
  N.0,
  N.1,
  d,
  gamma.c,
  alpha0.t,
  nu0.t,
  alpha1.t,
  nu1.t,
  s,
  m,
  design2 = NULL
)
}
\arguments{
\item{N.0}{number of subjects plan to be enrolled in control arm}

\item{N.1}{number of subjects plan to be enrolled in experimental arm}

\item{d}{expected number of events observed at time \code{l}}

\item{gamma.c}{parameter of the exponential distribution of censoring time}

\item{alpha0.t}{shape parameter of weibull survival distribution for control arm}

\item{nu0.t}{scale parameter of weibull survival distribution for control arm}

\item{alpha1.t}{shape parameters of weibull survival distribution for experimental arm}

\item{nu1.t}{scale parameter of a weibull survival distribution for control arm}

\item{s}{enrollment time}

\item{m}{maximum follow-up time for a subject}

\item{design2}{a list containing all the above parameters for two-arm design}
}
\description{
Title
}
