% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimData.R
\name{SimData}
\alias{SimData}
\title{Title}
\usage{
SimData(
  N = NULL,
  d = NULL,
  l = NULL,
  gamma = NULL,
  s = NULL,
  m = NULL,
  alpha = NULL,
  nu = NULL,
  design1,
  seed,
  nsim
)
}
\arguments{
\item{design1}{list containing all the above parameters for single-arm design}

\item{seed}{random seed}

\item{nsim}{number of simulations}
}
\description{
Title
}
