% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrialPred.TwoArm.R
\name{TrialPred.TwoArm}
\alias{TrialPred.TwoArm}
\title{Title}
\usage{
TrialPred.TwoArm(
  N.0 = NULL,
  N.1 = NULL,
  ratio = NULL,
  d = NULL,
  l = NULL,
  gamma.c = NULL,
  alpha0.t = NULL,
  nu0.t = NULL,
  alpha1.t = NULL,
  nu1.t = NULL,
  s = NULL,
  m = NULL
)
}
\arguments{
\item{m}{}
}
\description{
Title
}
