% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimData.TwoArm.R
\name{SimData.TwoArm}
\alias{SimData.TwoArm}
\title{Title}
\usage{
SimData.TwoArm(
  N.0 = NULL,
  N.1 = NULL,
  ratio = NULL,
  d = NULL,
  l = NULL,
  gamma.c = NULL,
  s = NULL,
  m = NULL,
  alpha0.t = NULL,
  nu0.t = NULL,
  HR = NULL,
  alpha1.t = NULL,
  nu1.t = NULL,
  design2 = NULL,
  seed = NULL,
  nsim = NULL
)
}
\arguments{
\item{design2}{list containing all the above parameters for two-arm design}

\item{seed}{random seed}

\item{nsim}{number of simulations}
}
\description{
Title
}
