% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrialPred.OneArm.R
\name{TrialPred.OneArm}
\alias{TrialPred.OneArm}
\title{Title}
\usage{
TrialPred.OneArm(
  N = NULL,
  d = NULL,
  l = NULL,
  gamma = NULL,
  s = NULL,
  m = NULL,
  alpha = NULL,
  nu = NULL,
  design1 = NULL
)
}
\arguments{
\item{N}{example}

\item{d}{example}

\item{l}{example}

\item{gamma}{example}

\item{s}{example}

\item{m}{example}

\item{alpha}{example}

\item{nu}{example}

\item{design1}{example}
}
\value{
example
}
\description{
Title
}
\examples{
ex
}
