% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrialPred.TwoArm.R
\name{TrialPred.TwoArm}
\alias{TrialPred.TwoArm}
\title{Title}
\usage{
TrialPred.TwoArm(
  N.0 = NULL,
  N.1 = NULL,
  ratio = NULL,
  d = NULL,
  l = NULL,
  gamma.c = NULL,
  alpha0.t = NULL,
  nu0.t = NULL,
  HR = NULL,
  alpha1.t = NULL,
  nu1.t = NULL,
  s = NULL,
  m = NULL
)
}
\arguments{
\item{N.0}{s}

\item{N.1}{s}

\item{ratio}{s}

\item{d}{s}

\item{l}{s}

\item{gamma.c}{s}

\item{alpha0.t}{s}

\item{nu0.t}{s}

\item{HR}{s}

\item{alpha1.t}{s}

\item{nu1.t}{s}

\item{s}{s}

\item{m}{s}
}
\value{
ssss
}
\description{
Title
}
\examples{
sssss
}
