% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrialPred.TwoArm.R
\name{TrialPred.TwoArm}
\alias{TrialPred.TwoArm}
\title{Title Predict two-arm clinical trial}
\usage{
TrialPred.TwoArm(
  N.0 = NULL,
  N.1 = NULL,
  ratio = NULL,
  d = NULL,
  l = NULL,
  gamma.c = NULL,
  alpha0.t = NULL,
  nu0.t = NULL,
  HR = NULL,
  alpha1.t = NULL,
  nu1.t = NULL,
  s = NULL,
  m = NULL
)
}
\arguments{
\item{N.0}{number of subjects plan to be enrolled in control arm}

\item{N.1}{number of subjects plan to be enrolled in experimental arm}

\item{ratio}{randomization ratio between two arms: N.1 / N.0}

\item{d}{expected number of events observed at time l}

\item{l}{observation time}

\item{gamma.c}{parameter of the exponential distribution of censoring time}

\item{alpha0.t}{shape parameter of weibull survival distribution for control arm}

\item{nu0.t}{scale parameter of weibull survival distribution for control arm}

\item{HR}{hazard ratio of experimental group over control group}

\item{alpha1.t}{shape parameters of weibull survival distribution for experimental arm}

\item{nu1.t}{scale parameter of a weibull survival distribution for control arm}

\item{s}{enrollment time}

\item{m}{maximum follow-up time for a subject}
}
\value{
This function returns a list including all design parameters as the same with input parameters of this function. If parameter \code{d} is missing, this function will calculate l, based on other parameters. If parameter N.0 or N.1 is missing, this function will calculate N.0 or N.1, based on other parameters. If parameter l is missing, this function will calculate l based on other parameters. If parameter gamma.c is missing, this function will calculate gamma.c based on other parameters.
}
\description{
Title Predict two-arm clinical trial
}
\examples{
# calculate the expected number of events
TrialPred.TwoArm(N.0=100,N.1=100,alpha0.t = 1,nu0.t=5,alpha1.t=2,nu1.t=4,gamma.c=1,s=5,m=4,l=6)
}
