% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrialPred.TwoArm.R
\name{TrialPred.TwoArm}
\alias{TrialPred.TwoArm}
\title{Title}
\usage{
TrialPred.TwoArm(
  N.0 = NULL,
  N.1 = NULL,
  ratio = NULL,
  d = NULL,
  l = NULL,
  gamma.c = NULL,
  alpha0.t = NULL,
  nu0.t = NULL,
  HR = NULL,
  alpha1.t = NULL,
  nu1.t = NULL,
  s = NULL,
  m = NULL
)
}
\arguments{
\item{N.0}{The number of subjects plan to be enrolled in control arm}

\item{N.1}{The number of subjects plan to be enrolled in experimental arm}

\item{ratio}{The Randomization ratio between two arms: N.1 / N.0}

\item{d}{The expected number of events observed at time l}

\item{l}{The observation time}

\item{gamma.c}{The parameter of the exponential distribution of censoring time}

\item{alpha0.t}{The shape parameter of weibull survival distribution for control arm}

\item{nu0.t}{The scale parameter of weibull survival distribution for control arm}

\item{HR}{The hazard ratio of experimental group over control group}

\item{alpha1.t}{The shape parameters of weibull survival distribution for experimental arm}

\item{nu1.t}{The scale parameter of a weibull survival distribution for control arm}

\item{s}{The enrollment time}

\item{m}{The maximum follow-up time for a subject}
}
\value{
ssss
}
\description{
Title
}
\examples{
sssss
}
